requirements.txt
# =========================================
# ULTIMATE AI/ML DEPENDENCIES FOR TRADING AI
# =========================================

# Core Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0
polars>=0.20.0
sympy>=1.12.0

# Deep Learning Frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
tensorflow>=2.14.0
jax[cpu]>=0.4.0

# Traditional Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Statistical Modeling
statsmodels>=0.14.0
imbalanced-learn>=0.11.0
category-encoders>=2.6.0

# Advanced ML & Optimization
optuna>=3.4.0
hyperopt>=0.2.7
ray[all]>=2.8.0
dask>=2023.12.0

# Time Series Analysis
sktime>=0.25.0
darts>=0.27.0
tsfresh>=0.20.0
prophet>=1.1.0

# Transformers & Modern DL
transformers>=4.35.0
sentence-transformers>=2.2.0
diffusers>=0.24.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.1.0
albumentations>=1.3.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.0
gensim>=4.3.0

# Graph Neural Networks
torch-geometric>=2.4.0
dgl>=2.0.0

# Reinforcement Learning
stable-baselines3>=2.1.0
gymnasium>=0.29.0
gym>=0.29.0

# Probabilistic Programming
pyro-ppl>=1.8.0
pymc>=5.7.0

# Model Interpretability
shap>=0.43.0
lime>=0.2.0
interpret>=0.5.0

# AutoML
flaml>=2.1.0
pycaret>=3.2.0

# Feature Engineering
featuretools>=1.28.0

# Visualization & Monitoring
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
wandb>=0.16.0
mlflow>=2.9.0

# MLOps & Deployment
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
gradio>=4.7.0

# Scientific Computing
numba>=0.58.0
joblib>=1.3.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
hypothesis>=6.87.0
black>=23.11.0
flake8>=6.1.0
coverage>=7.2.0

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.65.0

# =========================================
# TOPOLOGICAL DATA ANALYSIS (TDA) LIBRARIES
# =========================================
ripser>=0.6.4          # Persistent Homology computation (fast, optimized)
persim>=0.3.1          # Persistence diagram comparison (Wasserstein distance)
gudhi>=3.9.0           # Comprehensive TDA library (Persistent Homology, Mapper, etc.)
teaspoon>=1.0.0        # Time series analysis with TDA (financial time series)
pyts>=0.12.0           # Time series transformation (Recurrence Plot, Gramian Angular Field)
kmapper>=2.0.1         # Mapper algorithm for pattern discovery

# =========================================
# GPU COMPUTING (Optional - requires CUDA)
# =========================================
#cupy-cuda12x>=12.2.0  # Uncomment for GPU computing


